requirements.txt
Flask-Login~=0.6.3
Flask~=3.1.0
SQLAlchemy~=2.0.38
WTForms~=3.2.1
Werkzeug~=3.1.3
email-validator~=2.2.0
pillow~=11.2.1
transliterate~=1.10.2
ExifRead~=3.3.0
requests~=2.32.3
python-dotenv~=1.0.1